Cleanup and refactors for named ranges.
