+++ EPPlus/FormulaParsing/DependencyChain/DependencyChainFactory.cs
@@ -85,31 +85,6 @@ namespace OfficeOpenXml.FormulaParsing
 		#endregion
 
 		#region Private Static Methods
-		private static void GetChain(DependencyChain depChain, ILexer lexer, ExcelNamedRange name, ExcelCalculationOption options)
-		{
-			var ws = name.LocalSheet;
-			var id = ExcelCellBase.GetCellID(ws == null ? 0 : ws.SheetID, name.Index, 0);
-			if (!depChain.Index.ContainsKey(id))
-			{
-				// TODO: Need to update named range formula resolution here.
-				var f = new FormulaCell() { SheetID = ws == null ? -1 : ws.SheetID, Row = name.Index, Column = 0, Formula = name.NameFormula };
-				if (!string.IsNullOrEmpty(f.Formula))
-				{
-					f.Tokens = lexer.Tokenize(f.Formula, (ws == null ? null : ws.Name)).ToList();
-					if (ws == null)
-					{
-						name.Workbook.FormulaTokens.SetValue(name.Index, 0, f.Tokens);
-					}
-					else
-					{
-						ws._formulaTokens.SetValue(name.Index, 0, f.Tokens);
-					}
-					depChain.Add(f);
-					FollowChain(depChain, lexer, name.Workbook, ws, f, options);
-				}
-			}
-		}
-
 		private static void GetChain(DependencyChain depChain, ILexer lexer, ExcelWorksheet ws, string formula, ExcelCalculationOption options)
 		{
 			var f = new FormulaCell() { SheetID = ws.SheetID, Row = -1, Column = -1 };
