+++ EPPlusTest/ExcelNamedRangeCollectionTest.cs
@@ -1,5 +1,4 @@
-﻿using System.IO;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 using OfficeOpenXml;
 
 namespace EPPlusTest
@@ -7,822 +6,6 @@ namespace EPPlusTest
 	[TestClass]
 	public class ExcelNamedRangeCollectionTest
 	{
-		#region Named Range Integration Tests
-		#region Copy/Delete Worksheet Tests
-		[TestMethod]
-		public void CopyWorksheetWithWorksheetScopedNamedRanges()
-		{
-			var tempFile = new FileInfo(Path.GetTempFileName());
-			if (tempFile.Exists)
-				tempFile.Delete();
-			try
-			{
-				using (var excelPackage = new ExcelPackage())
-				{
-					var sheet1 = excelPackage.Workbook.Worksheets.Add("Sheet1");
-					var sheet2 = excelPackage.Workbook.Worksheets.Add("Sheet2");
-					sheet1.Names.Add("name1", "CONCATENATE(Sheet1!B2, Sheet1!$B$2)");
-					sheet1.Names.Add("name2", "CONCATENATE(Sheet2!B2, Sheet2!$B$2)");
-					sheet2.Names.Add("name3", "CONCATENATE(Sheet1!B2, Sheet1!$B$2)");
-					var sheet1Copy = excelPackage.Workbook.Worksheets.Copy("Sheet1", "Sheet1 copy");
-					Assert.AreEqual(0, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual(2, sheet1.Names.Count);
-					Assert.AreEqual(2, sheet1Copy.Names.Count);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE(Sheet1!B2, Sheet1!$B$2)", sheet1.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE(Sheet2!B2, Sheet2!$B$2)", sheet1.Names["name2"].NameFormula);
-					Assert.AreEqual("CONCATENATE(Sheet1!B2, Sheet1!$B$2)", sheet2.Names["name3"].NameFormula);
-					Assert.AreEqual("CONCATENATE('Sheet1 copy'!B2,'Sheet1 copy'!$B$2)", sheet1Copy.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE('SHEET2'!B2,'SHEET2'!$B$2)", sheet1Copy.Names["name2"].NameFormula);
-					excelPackage.SaveAs(tempFile);
-				}
-				using (var excelPackage = new ExcelPackage(tempFile))
-				{
-					var sheet1 = excelPackage.Workbook.Worksheets["Sheet1"];
-					var sheet1Copy = excelPackage.Workbook.Worksheets["Sheet1 copy"];
-					var sheet2 = excelPackage.Workbook.Worksheets["Sheet2"];
-					Assert.AreEqual(0, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual(2, sheet1.Names.Count);
-					Assert.AreEqual(2, sheet1Copy.Names.Count);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE(Sheet1!B2, Sheet1!$B$2)", sheet1.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE(Sheet2!B2, Sheet2!$B$2)", sheet1.Names["name2"].NameFormula);
-					Assert.AreEqual("CONCATENATE(Sheet1!B2, Sheet1!$B$2)", sheet2.Names["name3"].NameFormula);
-					Assert.AreEqual("CONCATENATE('Sheet1 copy'!B2,'Sheet1 copy'!$B$2)", sheet1Copy.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE('SHEET2'!B2,'SHEET2'!$B$2)", sheet1Copy.Names["name2"].NameFormula);
-				}
-			}
-			finally
-			{
-				if (tempFile.Exists)
-					tempFile.Delete();
-			}
-		}
-
-		[TestMethod]
-		public void CopyWorksheetWithWorkbookScopedNamedRanges()
-		{
-			var tempFile = new FileInfo(Path.GetTempFileName());
-			if (tempFile.Exists)
-				tempFile.Delete();
-			try
-			{
-				using (var excelPackage = new ExcelPackage())
-				{
-					var sheet1 = excelPackage.Workbook.Worksheets.Add("Sheet1");
-					var sheet2 = excelPackage.Workbook.Worksheets.Add("Sheet2");
-					excelPackage.Workbook.Names.Add("name1", "CONCATENATE(Sheet1!B2, Sheet1!$B$2)");
-					excelPackage.Workbook.Names.Add("name2", "CONCATENATE(Sheet2!B2, Sheet2!$B$2)");
-					var sheet1Copy = excelPackage.Workbook.Worksheets.Copy("Sheet1", "Sheet1 copy");
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual(0, sheet1.Names.Count);
-					Assert.AreEqual(0, sheet1Copy.Names.Count);
-					Assert.AreEqual(0, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE(Sheet1!B2, Sheet1!$B$2)", excelPackage.Workbook.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE(Sheet2!B2, Sheet2!$B$2)", excelPackage.Workbook.Names["name2"].NameFormula);
-					excelPackage.SaveAs(tempFile);
-				}
-				using (var excelPackage = new ExcelPackage(tempFile))
-				{
-					var sheet1 = excelPackage.Workbook.Worksheets["Sheet1"];
-					var sheet1Copy = excelPackage.Workbook.Worksheets["Sheet1 copy"];
-					var sheet2 = excelPackage.Workbook.Worksheets["Sheet2"];
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual(0, sheet1.Names.Count);
-					Assert.AreEqual(0, sheet1Copy.Names.Count);
-					Assert.AreEqual(0, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE(Sheet1!B2, Sheet1!$B$2)", excelPackage.Workbook.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE(Sheet2!B2, Sheet2!$B$2)", excelPackage.Workbook.Names["name2"].NameFormula);
-				}
-			}
-			finally
-			{
-				if (tempFile.Exists)
-					tempFile.Delete();
-			}
-		}
-
-		[TestMethod]
-		public void DeleteWorksheetWithWorksheetScopedNamedRanges()
-		{
-			var tempFile = new FileInfo(Path.GetTempFileName());
-			if (tempFile.Exists)
-				tempFile.Delete();
-			try
-			{
-				using (var excelPackage = new ExcelPackage())
-				{
-					var sheet1 = excelPackage.Workbook.Worksheets.Add("Sheet1");
-					var sheet2 = excelPackage.Workbook.Worksheets.Add("Sheet2");
-					sheet1.Names.Add("name1", "CONCATENATE(Sheet1!B2, Sheet1!$B$2)");
-					sheet1.Names.Add("name2", "CONCATENATE(Sheet2!B2, Sheet2!$B$2)");
-					sheet2.Names.Add("name3", "CONCATENATE(Sheet1!B2, Sheet1!$B$2)");
-					sheet2.Names.Add("name4", "CONCATENATE(Sheet2!B2, Sheet2!$B$2)");
-					excelPackage.Workbook.Worksheets.Delete(sheet1);
-					Assert.AreEqual(0, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual(2, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE(#REF!B2,#REF!$B$2)", sheet2.Names["name3"].NameFormula);
-					Assert.AreEqual("CONCATENATE('SHEET2'!B2,'SHEET2'!$B$2)", sheet2.Names["name4"].NameFormula);
-					excelPackage.SaveAs(tempFile);
-				}
-				using (var excelPackage = new ExcelPackage(tempFile))
-				{
-					var sheet2 = excelPackage.Workbook.Worksheets["Sheet2"];
-					Assert.AreEqual(2, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE(#REF!B2,#REF!$B$2)", sheet2.Names["name3"].NameFormula);
-					Assert.AreEqual("CONCATENATE('SHEET2'!B2,'SHEET2'!$B$2)", sheet2.Names["name4"].NameFormula);
-				}
-			}
-			finally
-			{
-				if (tempFile.Exists)
-					tempFile.Delete();
-			}
-		}
-
-		[TestMethod]
-		public void DeleteWorksheetWithWorkbookScopedNamedRanges()
-		{
-			var tempFile = new FileInfo(Path.GetTempFileName());
-			if (tempFile.Exists)
-				tempFile.Delete();
-			try
-			{
-				using (var excelPackage = new ExcelPackage())
-				{
-					var sheet1 = excelPackage.Workbook.Worksheets.Add("Sheet1");
-					var sheet2 = excelPackage.Workbook.Worksheets.Add("Sheet2");
-					excelPackage.Workbook.Names.Add("name1", "CONCATENATE(Sheet1!B2, Sheet1!$B$2)");
-					excelPackage.Workbook.Names.Add("name2", "CONCATENATE(Sheet2!B2, Sheet2!$B$2)");
-					excelPackage.Workbook.Worksheets.Delete(sheet1);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual(0, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE(#REF!B2,#REF!$B$2)", excelPackage.Workbook.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE('SHEET2'!B2,'SHEET2'!$B$2)", excelPackage.Workbook.Names["name2"].NameFormula);
-					excelPackage.SaveAs(tempFile);
-				}
-				using (var excelPackage = new ExcelPackage(tempFile))
-				{
-					var sheet2 = excelPackage.Workbook.Worksheets["Sheet2"];
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual(0, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE(#REF!B2,#REF!$B$2)", excelPackage.Workbook.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE('SHEET2'!B2,'SHEET2'!$B$2)", excelPackage.Workbook.Names["name2"].NameFormula);
-				}
-			}
-			finally
-			{
-				if (tempFile.Exists)
-					tempFile.Delete();
-			}
-		}
-		#endregion
-
-		#region Insert Row(s) Tests
-		[TestMethod]
-		public void InsertRowBeforeNamedRanges()
-		{
-			var tempFile = new FileInfo(Path.GetTempFileName());
-			if (tempFile.Exists)
-				tempFile.Delete();
-			try
-			{
-				using (var excelPackage = new ExcelPackage())
-				{
-					var sheet1 = excelPackage.Workbook.Worksheets.Add("Sheet1");
-					var sheet2 = excelPackage.Workbook.Worksheets.Add("Sheet2");
-					sheet1.Names.Add("name1", "CONCATENATE(Sheet1!B5, Sheet1!$B$5)");
-					sheet1.Names.Add("name2", "CONCATENATE(Sheet2!B5, Sheet2!$B$5)");
-					sheet2.Names.Add("name3", "CONCATENATE(Sheet1!B5, Sheet1!$B$5)");
-					excelPackage.Workbook.Names.Add("name4", "CONCATENATE(Sheet1!B5, Sheet1!$B$5)");
-					excelPackage.Workbook.Names.Add("name5", "CONCATENATE(Sheet2!B5, Sheet2!$B$5)");
-					sheet1.InsertRow(4, 1);
-					Assert.AreEqual(2, sheet1.Names.Count);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$6)", sheet1.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!B5,SHEET2!$B$5)", sheet1.Names["name2"].NameFormula);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$6)", sheet2.Names["name3"].NameFormula);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$6)", excelPackage.Workbook.Names["name4"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!B5,SHEET2!$B$5)", excelPackage.Workbook.Names["name5"].NameFormula);
-					excelPackage.SaveAs(tempFile);
-				}
-				using (var excelPackage = new ExcelPackage(tempFile))
-				{
-					var sheet1 = excelPackage.Workbook.Worksheets["Sheet1"];
-					var sheet2 = excelPackage.Workbook.Worksheets["Sheet2"];
-					Assert.AreEqual(2, sheet1.Names.Count);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$6)", sheet1.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!B5,SHEET2!$B$5)", sheet1.Names["name2"].NameFormula);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$6)", sheet2.Names["name3"].NameFormula);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$6)", excelPackage.Workbook.Names["name4"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!B5,SHEET2!$B$5)", excelPackage.Workbook.Names["name5"].NameFormula);
-				}
-			}
-			finally
-			{
-				if (tempFile.Exists)
-					tempFile.Delete();
-			}
-		}
-
-		[TestMethod]
-		public void InsertMultipleRowsBeforeNamedRanges()
-		{
-			var tempFile = new FileInfo(Path.GetTempFileName());
-			if (tempFile.Exists)
-				tempFile.Delete();
-			try
-			{
-				using (var excelPackage = new ExcelPackage())
-				{
-					var sheet1 = excelPackage.Workbook.Worksheets.Add("Sheet1");
-					var sheet2 = excelPackage.Workbook.Worksheets.Add("Sheet2");
-					sheet1.Names.Add("name1", "CONCATENATE(Sheet1!B5, Sheet1!$B$5)");
-					sheet1.Names.Add("name2", "CONCATENATE(Sheet2!B5, Sheet2!$B$5)");
-					sheet2.Names.Add("name3", "CONCATENATE(Sheet1!B5, Sheet1!$B$5)");
-					excelPackage.Workbook.Names.Add("name4", "CONCATENATE(Sheet1!B5, Sheet1!$B$5)");
-					excelPackage.Workbook.Names.Add("name5", "CONCATENATE(Sheet2!B5, Sheet2!$B$5)");
-					sheet1.InsertRow(4, 3);
-					Assert.AreEqual(2, sheet1.Names.Count);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$8)", sheet1.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!B5,SHEET2!$B$5)", sheet1.Names["name2"].NameFormula);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$8)", sheet2.Names["name3"].NameFormula);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$8)", excelPackage.Workbook.Names["name4"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!B5,SHEET2!$B$5)", excelPackage.Workbook.Names["name5"].NameFormula);
-					excelPackage.SaveAs(tempFile);
-				}
-				using (var excelPackage = new ExcelPackage(tempFile))
-				{
-					var sheet1 = excelPackage.Workbook.Worksheets["Sheet1"];
-					var sheet2 = excelPackage.Workbook.Worksheets["Sheet2"];
-					Assert.AreEqual(2, sheet1.Names.Count);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$8)", sheet1.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!B5,SHEET2!$B$5)", sheet1.Names["name2"].NameFormula);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$8)", sheet2.Names["name3"].NameFormula);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$8)", excelPackage.Workbook.Names["name4"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!B5,SHEET2!$B$5)", excelPackage.Workbook.Names["name5"].NameFormula);
-				}
-			}
-			finally
-			{
-				if (tempFile.Exists)
-					tempFile.Delete();
-			}
-		}
-
-		[TestMethod]
-		public void InsertRowAfterNamedRanges()
-		{
-			var tempFile = new FileInfo(Path.GetTempFileName());
-			if (tempFile.Exists)
-				tempFile.Delete();
-			try
-			{
-				using (var excelPackage = new ExcelPackage())
-				{
-					var sheet1 = excelPackage.Workbook.Worksheets.Add("Sheet1");
-					var sheet2 = excelPackage.Workbook.Worksheets.Add("Sheet2");
-					sheet1.Names.Add("name1", "CONCATENATE(Sheet1!B5, Sheet1!$B$5)");
-					sheet1.Names.Add("name2", "CONCATENATE(Sheet2!B5, Sheet2!$B$5)");
-					sheet2.Names.Add("name3", "CONCATENATE(Sheet1!B5, Sheet1!$B$5)");
-					excelPackage.Workbook.Names.Add("name4", "CONCATENATE(Sheet1!B5, Sheet1!$B$5)");
-					excelPackage.Workbook.Names.Add("name5", "CONCATENATE(Sheet2!B5, Sheet2!$B$5)");
-					sheet1.InsertRow(6, 1);
-					Assert.AreEqual(2, sheet1.Names.Count);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$5)", sheet1.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!B5,SHEET2!$B$5)", sheet1.Names["name2"].NameFormula);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$5)", sheet2.Names["name3"].NameFormula);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$5)", excelPackage.Workbook.Names["name4"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!B5,SHEET2!$B$5)", excelPackage.Workbook.Names["name5"].NameFormula);
-					excelPackage.SaveAs(tempFile);
-				}
-				using (var excelPackage = new ExcelPackage(tempFile))
-				{
-					var sheet1 = excelPackage.Workbook.Worksheets["Sheet1"];
-					var sheet2 = excelPackage.Workbook.Worksheets["Sheet2"];
-					Assert.AreEqual(2, sheet1.Names.Count);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$5)", sheet1.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!B5,SHEET2!$B$5)", sheet1.Names["name2"].NameFormula);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$5)", sheet2.Names["name3"].NameFormula);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$5)", excelPackage.Workbook.Names["name4"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!B5,SHEET2!$B$5)", excelPackage.Workbook.Names["name5"].NameFormula);
-				}
-			}
-			finally
-			{
-				if (tempFile.Exists)
-					tempFile.Delete();
-			}
-		}
-		#endregion
-
-		#region Insert Column(s) Tests
-		[TestMethod]
-		public void InsertColumnBeforeNamedRanges()
-		{
-			var tempFile = new FileInfo(Path.GetTempFileName());
-			if (tempFile.Exists)
-				tempFile.Delete();
-			try
-			{
-				using (var excelPackage = new ExcelPackage())
-				{
-					var sheet1 = excelPackage.Workbook.Worksheets.Add("Sheet1");
-					var sheet2 = excelPackage.Workbook.Worksheets.Add("Sheet2");
-					sheet1.Names.Add("name1", "CONCATENATE(Sheet1!D5, Sheet1!$D$5)");
-					sheet1.Names.Add("name2", "CONCATENATE(Sheet2!D5, Sheet2!$D$5)");
-					sheet2.Names.Add("name3", "CONCATENATE(Sheet1!D5, Sheet1!$D$5)");
-					excelPackage.Workbook.Names.Add("name4", "CONCATENATE(Sheet1!D5, Sheet1!$D$5)");
-					excelPackage.Workbook.Names.Add("name5", "CONCATENATE(Sheet2!D5, Sheet2!$D$5)");
-					sheet1.InsertColumn(2, 1);
-					Assert.AreEqual(2, sheet1.Names.Count);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!D5,'SHEET1'!$E$5)", sheet1.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!D5,SHEET2!$D$5)", sheet1.Names["name2"].NameFormula);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!D5,'SHEET1'!$E$5)", sheet2.Names["name3"].NameFormula);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!D5,'SHEET1'!$E$5)", excelPackage.Workbook.Names["name4"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!D5,SHEET2!$D$5)", excelPackage.Workbook.Names["name5"].NameFormula);
-					excelPackage.SaveAs(tempFile);
-				}
-				using (var excelPackage = new ExcelPackage(tempFile))
-				{
-					var sheet1 = excelPackage.Workbook.Worksheets["Sheet1"];
-					var sheet2 = excelPackage.Workbook.Worksheets["Sheet2"];
-					Assert.AreEqual(2, sheet1.Names.Count);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!D5,'SHEET1'!$E$5)", sheet1.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!D5,SHEET2!$D$5)", sheet1.Names["name2"].NameFormula);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!D5,'SHEET1'!$E$5)", sheet2.Names["name3"].NameFormula);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!D5,'SHEET1'!$E$5)", excelPackage.Workbook.Names["name4"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!D5,SHEET2!$D$5)", excelPackage.Workbook.Names["name5"].NameFormula);
-				}
-			}
-			finally
-			{
-				if (tempFile.Exists)
-					tempFile.Delete();
-			}
-		}
-
-		[TestMethod]
-		public void InsertMultipleColumnsBeforeNamedRanges()
-		{
-			var tempFile = new FileInfo(Path.GetTempFileName());
-			if (tempFile.Exists)
-				tempFile.Delete();
-			try
-			{
-				using (var excelPackage = new ExcelPackage())
-				{
-					var sheet1 = excelPackage.Workbook.Worksheets.Add("Sheet1");
-					var sheet2 = excelPackage.Workbook.Worksheets.Add("Sheet2");
-					sheet1.Names.Add("name1", "CONCATENATE(Sheet1!D5, Sheet1!$D$5)");
-					sheet1.Names.Add("name2", "CONCATENATE(Sheet2!D5, Sheet2!$D$5)");
-					sheet2.Names.Add("name3", "CONCATENATE(Sheet1!D5, Sheet1!$D$5)");
-					excelPackage.Workbook.Names.Add("name4", "CONCATENATE(Sheet1!D5, Sheet1!$D$5)");
-					excelPackage.Workbook.Names.Add("name5", "CONCATENATE(Sheet2!D5, Sheet2!$D$5)");
-					sheet1.InsertColumn(2, 3);
-					Assert.AreEqual(2, sheet1.Names.Count);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!D5,'SHEET1'!$G$5)", sheet1.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!D5,SHEET2!$D$5)", sheet1.Names["name2"].NameFormula);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!D5,'SHEET1'!$G$5)", sheet2.Names["name3"].NameFormula);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!D5,'SHEET1'!$G$5)", excelPackage.Workbook.Names["name4"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!D5,SHEET2!$D$5)", excelPackage.Workbook.Names["name5"].NameFormula);
-					excelPackage.SaveAs(tempFile);
-				}
-				using (var excelPackage = new ExcelPackage(tempFile))
-				{
-					var sheet1 = excelPackage.Workbook.Worksheets["Sheet1"];
-					var sheet2 = excelPackage.Workbook.Worksheets["Sheet2"];
-					Assert.AreEqual(2, sheet1.Names.Count);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!D5,'SHEET1'!$G$5)", sheet1.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!D5,SHEET2!$D$5)", sheet1.Names["name2"].NameFormula);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!D5,'SHEET1'!$G$5)", sheet2.Names["name3"].NameFormula);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!D5,'SHEET1'!$G$5)", excelPackage.Workbook.Names["name4"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!D5,SHEET2!$D$5)", excelPackage.Workbook.Names["name5"].NameFormula);
-				}
-			}
-			finally
-			{
-				if (tempFile.Exists)
-					tempFile.Delete();
-			}
-		}
-
-		[TestMethod]
-		public void InsertColumnAfterNamedRanges()
-		{
-			var tempFile = new FileInfo(Path.GetTempFileName());
-			if (tempFile.Exists)
-				tempFile.Delete();
-			try
-			{
-				using (var excelPackage = new ExcelPackage())
-				{
-					var sheet1 = excelPackage.Workbook.Worksheets.Add("Sheet1");
-					var sheet2 = excelPackage.Workbook.Worksheets.Add("Sheet2");
-					sheet1.Names.Add("name1", "CONCATENATE(Sheet1!B5, Sheet1!$B$5)");
-					sheet1.Names.Add("name2", "CONCATENATE(Sheet2!B5, Sheet2!$B$5)");
-					sheet2.Names.Add("name3", "CONCATENATE(Sheet1!B5, Sheet1!$B$5)");
-					excelPackage.Workbook.Names.Add("name4", "CONCATENATE(Sheet1!B5, Sheet1!$B$5)");
-					excelPackage.Workbook.Names.Add("name5", "CONCATENATE(Sheet2!B5, Sheet2!$B$5)");
-					sheet1.InsertColumn(6, 1);
-					Assert.AreEqual(2, sheet1.Names.Count);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$5)", sheet1.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!B5,SHEET2!$B$5)", sheet1.Names["name2"].NameFormula);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$5)", sheet2.Names["name3"].NameFormula);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$5)", excelPackage.Workbook.Names["name4"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!B5,SHEET2!$B$5)", excelPackage.Workbook.Names["name5"].NameFormula);
-					excelPackage.SaveAs(tempFile);
-				}
-				using (var excelPackage = new ExcelPackage(tempFile))
-				{
-					var sheet1 = excelPackage.Workbook.Worksheets["Sheet1"];
-					var sheet2 = excelPackage.Workbook.Worksheets["Sheet2"];
-					Assert.AreEqual(2, sheet1.Names.Count);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$5)", sheet1.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!B5,SHEET2!$B$5)", sheet1.Names["name2"].NameFormula);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$5)", sheet2.Names["name3"].NameFormula);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$5)", excelPackage.Workbook.Names["name4"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!B5,SHEET2!$B$5)", excelPackage.Workbook.Names["name5"].NameFormula);
-				}
-			}
-			finally
-			{
-				if (tempFile.Exists)
-					tempFile.Delete();
-			}
-		}
-		#endregion
-
-		#region Delete Row(s) Tests
-		[TestMethod]
-		public void DeleteRowBeforeNamedRanges()
-		{
-			var tempFile = new FileInfo(Path.GetTempFileName());
-			if (tempFile.Exists)
-				tempFile.Delete();
-			try
-			{
-				using (var excelPackage = new ExcelPackage())
-				{
-					var sheet1 = excelPackage.Workbook.Worksheets.Add("Sheet1");
-					var sheet2 = excelPackage.Workbook.Worksheets.Add("Sheet2");
-					sheet1.Names.Add("name1", "CONCATENATE(Sheet1!D5, Sheet1!$D$5)");
-					sheet1.Names.Add("name2", "CONCATENATE(Sheet2!D5, Sheet2!$D$5)");
-					sheet2.Names.Add("name3", "CONCATENATE(Sheet1!D5, Sheet1!$D$5)");
-					excelPackage.Workbook.Names.Add("name4", "CONCATENATE(Sheet1!D5, Sheet1!$D$5)");
-					excelPackage.Workbook.Names.Add("name5", "CONCATENATE(Sheet2!D5, Sheet2!$D$5)");
-					sheet1.DeleteRow(3, 1);
-					Assert.AreEqual(2, sheet1.Names.Count);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!D5,'SHEET1'!$D$4)", sheet1.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!D5,SHEET2!$D$5)", sheet1.Names["name2"].NameFormula);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!D5,'SHEET1'!$D$4)", sheet2.Names["name3"].NameFormula);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!D5,'SHEET1'!$D$4)", excelPackage.Workbook.Names["name4"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!D5,SHEET2!$D$5)", excelPackage.Workbook.Names["name5"].NameFormula);
-					excelPackage.SaveAs(tempFile);
-				}
-				using (var excelPackage = new ExcelPackage(tempFile))
-				{
-					var sheet1 = excelPackage.Workbook.Worksheets["Sheet1"];
-					var sheet2 = excelPackage.Workbook.Worksheets["Sheet2"];
-					Assert.AreEqual(2, sheet1.Names.Count);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!D5,'SHEET1'!$D$4)", sheet1.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!D5,SHEET2!$D$5)", sheet1.Names["name2"].NameFormula);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!D5,'SHEET1'!$D$4)", sheet2.Names["name3"].NameFormula);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!D5,'SHEET1'!$D$4)", excelPackage.Workbook.Names["name4"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!D5,SHEET2!$D$5)", excelPackage.Workbook.Names["name5"].NameFormula);
-				}
-			}
-			finally
-			{
-				if (tempFile.Exists)
-					tempFile.Delete();
-			}
-		}
-
-		[TestMethod]
-		public void DeleteMultipleRowsBeforeNamedRanges()
-		{
-			var tempFile = new FileInfo(Path.GetTempFileName());
-			if (tempFile.Exists)
-				tempFile.Delete();
-			try
-			{
-				using (var excelPackage = new ExcelPackage())
-				{
-					var sheet1 = excelPackage.Workbook.Worksheets.Add("Sheet1");
-					var sheet2 = excelPackage.Workbook.Worksheets.Add("Sheet2");
-					sheet1.Names.Add("name1", "CONCATENATE(Sheet1!F5, Sheet1!$F$5)");
-					sheet1.Names.Add("name2", "CONCATENATE(Sheet2!F5, Sheet2!$F$5)");
-					sheet2.Names.Add("name3", "CONCATENATE(Sheet1!F5, Sheet1!$F$5)");
-					excelPackage.Workbook.Names.Add("name4", "CONCATENATE(Sheet1!F5, Sheet1!$F$5)");
-					excelPackage.Workbook.Names.Add("name5", "CONCATENATE(Sheet2!F5, Sheet2!$F$5)");
-					sheet1.DeleteRow(2, 3);
-					Assert.AreEqual(2, sheet1.Names.Count);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!F5,'SHEET1'!$F$2)", sheet1.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!F5,SHEET2!$F$5)", sheet1.Names["name2"].NameFormula);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!F5,'SHEET1'!$F$2)", sheet2.Names["name3"].NameFormula);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!F5,'SHEET1'!$F$2)", excelPackage.Workbook.Names["name4"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!F5,SHEET2!$F$5)", excelPackage.Workbook.Names["name5"].NameFormula);
-					excelPackage.SaveAs(tempFile);
-				}
-				using (var excelPackage = new ExcelPackage(tempFile))
-				{
-					var sheet1 = excelPackage.Workbook.Worksheets["Sheet1"];
-					var sheet2 = excelPackage.Workbook.Worksheets["Sheet2"];
-					Assert.AreEqual(2, sheet1.Names.Count);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!F5,'SHEET1'!$F$2)", sheet1.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!F5,SHEET2!$F$5)", sheet1.Names["name2"].NameFormula);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!F5,'SHEET1'!$F$2)", sheet2.Names["name3"].NameFormula);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!F5,'SHEET1'!$F$2)", excelPackage.Workbook.Names["name4"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!F5,SHEET2!$F$5)", excelPackage.Workbook.Names["name5"].NameFormula);
-				}
-			}
-			finally
-			{
-				if (tempFile.Exists)
-					tempFile.Delete();
-			}
-		}
-
-		[TestMethod]
-		public void DeleteRowAfterNamedRanges()
-		{
-			var tempFile = new FileInfo(Path.GetTempFileName());
-			if (tempFile.Exists)
-				tempFile.Delete();
-			try
-			{
-				using (var excelPackage = new ExcelPackage())
-				{
-					var sheet1 = excelPackage.Workbook.Worksheets.Add("Sheet1");
-					var sheet2 = excelPackage.Workbook.Worksheets.Add("Sheet2");
-					sheet1.Names.Add("name1", "CONCATENATE(Sheet1!B5, Sheet1!$B$5)");
-					sheet1.Names.Add("name2", "CONCATENATE(Sheet2!B5, Sheet2!$B$5)");
-					sheet2.Names.Add("name3", "CONCATENATE(Sheet1!B5, Sheet1!$B$5)");
-					excelPackage.Workbook.Names.Add("name4", "CONCATENATE(Sheet1!B5, Sheet1!$B$5)");
-					excelPackage.Workbook.Names.Add("name5", "CONCATENATE(Sheet2!B5, Sheet2!$B$5)");
-					sheet1.DeleteRow(7, 1);
-					Assert.AreEqual(2, sheet1.Names.Count);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$5)", sheet1.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!B5,SHEET2!$B$5)", sheet1.Names["name2"].NameFormula);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$5)", sheet2.Names["name3"].NameFormula);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$5)", excelPackage.Workbook.Names["name4"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!B5,SHEET2!$B$5)", excelPackage.Workbook.Names["name5"].NameFormula);
-					excelPackage.SaveAs(tempFile);
-				}
-				using (var excelPackage = new ExcelPackage(tempFile))
-				{
-					var sheet1 = excelPackage.Workbook.Worksheets["Sheet1"];
-					var sheet2 = excelPackage.Workbook.Worksheets["Sheet2"];
-					Assert.AreEqual(2, sheet1.Names.Count);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$5)", sheet1.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!B5,SHEET2!$B$5)", sheet1.Names["name2"].NameFormula);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$5)", sheet2.Names["name3"].NameFormula);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$5)", excelPackage.Workbook.Names["name4"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!B5,SHEET2!$B$5)", excelPackage.Workbook.Names["name5"].NameFormula);
-				}
-			}
-			finally
-			{
-				if (tempFile.Exists)
-					tempFile.Delete();
-			}
-		}
-		#endregion
-
-		#region Delete Columnn(s) Tests
-		[TestMethod]
-		public void DeleteColumnBeforeNamedRanges()
-		{
-			var tempFile = new FileInfo(Path.GetTempFileName());
-			if (tempFile.Exists)
-				tempFile.Delete();
-			try
-			{
-				using (var excelPackage = new ExcelPackage())
-				{
-					var sheet1 = excelPackage.Workbook.Worksheets.Add("Sheet1");
-					var sheet2 = excelPackage.Workbook.Worksheets.Add("Sheet2");
-					sheet1.Names.Add("name1", "CONCATENATE(Sheet1!D5, Sheet1!$D$5)");
-					sheet1.Names.Add("name2", "CONCATENATE(Sheet2!D5, Sheet2!$D$5)");
-					sheet2.Names.Add("name3", "CONCATENATE(Sheet1!D5, Sheet1!$D$5)");
-					excelPackage.Workbook.Names.Add("name4", "CONCATENATE(Sheet1!D5, Sheet1!$D$5)");
-					excelPackage.Workbook.Names.Add("name5", "CONCATENATE(Sheet2!D5, Sheet2!$D$5)");
-					sheet1.DeleteColumn(2, 1);
-					Assert.AreEqual(2, sheet1.Names.Count);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!D5,'SHEET1'!$C$5)", sheet1.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!D5,SHEET2!$D$5)", sheet1.Names["name2"].NameFormula);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!D5,'SHEET1'!$C$5)", sheet2.Names["name3"].NameFormula);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!D5,'SHEET1'!$C$5)", excelPackage.Workbook.Names["name4"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!D5,SHEET2!$D$5)", excelPackage.Workbook.Names["name5"].NameFormula);
-					excelPackage.SaveAs(tempFile);
-				}
-				using (var excelPackage = new ExcelPackage(tempFile))
-				{
-					var sheet1 = excelPackage.Workbook.Worksheets["Sheet1"];
-					var sheet2 = excelPackage.Workbook.Worksheets["Sheet2"];
-					Assert.AreEqual(2, sheet1.Names.Count);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!D5,'SHEET1'!$C$5)", sheet1.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!D5,SHEET2!$D$5)", sheet1.Names["name2"].NameFormula);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!D5,'SHEET1'!$C$5)", sheet2.Names["name3"].NameFormula);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!D5,'SHEET1'!$C$5)", excelPackage.Workbook.Names["name4"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!D5,SHEET2!$D$5)", excelPackage.Workbook.Names["name5"].NameFormula);
-				}
-			}
-			finally
-			{
-				if (tempFile.Exists)
-					tempFile.Delete();
-			}
-		}
-
-		[TestMethod]
-		public void DeleteMultipleColumnsBeforeNamedRanges()
-		{
-			var tempFile = new FileInfo(Path.GetTempFileName());
-			if (tempFile.Exists)
-				tempFile.Delete();
-			try
-			{
-				using (var excelPackage = new ExcelPackage())
-				{
-					var sheet1 = excelPackage.Workbook.Worksheets.Add("Sheet1");
-					var sheet2 = excelPackage.Workbook.Worksheets.Add("Sheet2");
-					sheet1.Names.Add("name1", "CONCATENATE(Sheet1!F5, Sheet1!$F$5)");
-					sheet1.Names.Add("name2", "CONCATENATE(Sheet2!F5, Sheet2!$F$5)");
-					sheet2.Names.Add("name3", "CONCATENATE(Sheet1!F5, Sheet1!$F$5)");
-					excelPackage.Workbook.Names.Add("name4", "CONCATENATE(Sheet1!F5, Sheet1!$F$5)");
-					excelPackage.Workbook.Names.Add("name5", "CONCATENATE(Sheet2!F5, Sheet2!$F$5)");
-					sheet1.DeleteColumn(2, 3);
-					Assert.AreEqual(2, sheet1.Names.Count);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!F5,'SHEET1'!$C$5)", sheet1.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!F5,SHEET2!$F$5)", sheet1.Names["name2"].NameFormula);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!F5,'SHEET1'!$C$5)", sheet2.Names["name3"].NameFormula);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!F5,'SHEET1'!$C$5)", excelPackage.Workbook.Names["name4"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!F5,SHEET2!$F$5)", excelPackage.Workbook.Names["name5"].NameFormula);
-					excelPackage.SaveAs(tempFile);
-				}
-				using (var excelPackage = new ExcelPackage(tempFile))
-				{
-					var sheet1 = excelPackage.Workbook.Worksheets["Sheet1"];
-					var sheet2 = excelPackage.Workbook.Worksheets["Sheet2"];
-					Assert.AreEqual(2, sheet1.Names.Count);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!F5,'SHEET1'!$C$5)", sheet1.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!F5,SHEET2!$F$5)", sheet1.Names["name2"].NameFormula);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!F5,'SHEET1'!$C$5)", sheet2.Names["name3"].NameFormula);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!F5,'SHEET1'!$C$5)", excelPackage.Workbook.Names["name4"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!F5,SHEET2!$F$5)", excelPackage.Workbook.Names["name5"].NameFormula);
-				}
-			}
-			finally
-			{
-				if (tempFile.Exists)
-					tempFile.Delete();
-			}
-		}
-
-		[TestMethod]
-		public void DeleteColumnAfterNamedRanges()
-		{
-			var tempFile = new FileInfo(Path.GetTempFileName());
-			if (tempFile.Exists)
-				tempFile.Delete();
-			try
-			{
-				using (var excelPackage = new ExcelPackage())
-				{
-					var sheet1 = excelPackage.Workbook.Worksheets.Add("Sheet1");
-					var sheet2 = excelPackage.Workbook.Worksheets.Add("Sheet2");
-					sheet1.Names.Add("name1", "CONCATENATE(Sheet1!B5, Sheet1!$B$5)");
-					sheet1.Names.Add("name2", "CONCATENATE(Sheet2!B5, Sheet2!$B$5)");
-					sheet2.Names.Add("name3", "CONCATENATE(Sheet1!B5, Sheet1!$B$5)");
-					excelPackage.Workbook.Names.Add("name4", "CONCATENATE(Sheet1!B5, Sheet1!$B$5)");
-					excelPackage.Workbook.Names.Add("name5", "CONCATENATE(Sheet2!B5, Sheet2!$B$5)");
-					sheet1.InsertColumn(6, 1);
-					Assert.AreEqual(2, sheet1.Names.Count);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$5)", sheet1.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!B5,SHEET2!$B$5)", sheet1.Names["name2"].NameFormula);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$5)", sheet2.Names["name3"].NameFormula);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$5)", excelPackage.Workbook.Names["name4"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!B5,SHEET2!$B$5)", excelPackage.Workbook.Names["name5"].NameFormula);
-					excelPackage.SaveAs(tempFile);
-				}
-				using (var excelPackage = new ExcelPackage(tempFile))
-				{
-					var sheet1 = excelPackage.Workbook.Worksheets["Sheet1"];
-					var sheet2 = excelPackage.Workbook.Worksheets["Sheet2"];
-					Assert.AreEqual(2, sheet1.Names.Count);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$5)", sheet1.Names["name1"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!B5,SHEET2!$B$5)", sheet1.Names["name2"].NameFormula);
-					Assert.AreEqual(1, sheet2.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$5)", sheet2.Names["name3"].NameFormula);
-					Assert.AreEqual(2, excelPackage.Workbook.Names.Count);
-					Assert.AreEqual("CONCATENATE('SHEET1'!B5,'SHEET1'!$B$5)", excelPackage.Workbook.Names["name4"].NameFormula);
-					Assert.AreEqual("CONCATENATE(SHEET2!B5,SHEET2!$B$5)", excelPackage.Workbook.Names["name5"].NameFormula);
-				}
-			}
-			finally
-			{
-				if (tempFile.Exists)
-					tempFile.Delete();
-			}
-		}
-		#endregion
-		#endregion
-
 		#region Insert Tests
 		[TestMethod]
 		public void InsertRowsBeforeAbsoluteNamedRange()
